[ refactor ] Move `reallyFree` function to TypeChecking.Free.Reduce
